Allow for creating documentation that can link to external dependencies
This isn't the prettiest of code, nor approachs, but the goal of this
action so far is to get everything working "end-to-end" and then go
back and see what additional support we need in tooling to make the
process less brittle.

The code in this commit makes a lot of assumptions about the layout
of dependencies pulled by corral. It shouldn't be doing that. That
knowledge should be in corral and then we could either:

- use parts of corral as a library from a pony version of this
- have a corral command to get said information dumped in some fashion

or something similar. In the meantime, this works.

As usual, I apologize for my Python. It is what it is.

Closes #3