Stop versioning. Start tracking ponyc releases automatically. (#23)
We used to do versioning for the library-documentation-action but this lead to
a lot of extra work each time a new version of ponyc was released. This action
is considered feature complete and we don't intend to do any breaking updates
to it. It will be deprecated in the future and replaced by a pony
documentation command that ships with ponyc.